Move MethodSymbolResolver to PDBs namespace
