+++ tracer/src/Datadog.Trace/PDBs/MethodSymbolResolver.cs
@@ -13,7 +13,7 @@ using Datadog.Trace.Vendors.dnlib.DotNet;
 using Datadog.Trace.Vendors.dnlib.DotNet.Emit;
 using Datadog.Trace.Vendors.dnlib.DotNet.Pdb;
 
-namespace Datadog.Trace.Ci
+namespace Datadog.Trace.PDBs
 {
     internal class MethodSymbolResolver
     {
